feat: ab#1604 load module sbom from minio
