+++ application/src/main/resources/application.properties
@@ -50,6 +50,7 @@ springdoc.swagger-ui.csrf.enabled=true
 minio.user=minioadmin
 minio.password=minioadmin
 minio.endpoint=http://localhost:9000
+minio.sbombucket=sboms
 
 # logging.level.org.springframework.security=DEBUG
 # logging.level.org.springframework.web=DEBUG
