+++ rest/src/main/java/de/starwit/rest/controller/ModuleController.java
@@ -1,11 +1,16 @@
 package de.starwit.rest.controller;
 
+import java.io.IOException;
+import java.security.InvalidKeyException;
 import java.util.List;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -20,6 +25,8 @@ import org.springframework.web.bind.annotation.RestController;
 import de.starwit.persistence.entity.ModuleEntity;
 import de.starwit.persistence.exception.NotificationException;
 import de.starwit.rest.exception.NotificationDto;
+import de.starwit.service.impl.MinioException;
+import de.starwit.service.impl.MinioService;
 import de.starwit.service.impl.ModuleService;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.persistence.EntityNotFoundException;
@@ -34,13 +41,28 @@ public class ModuleController {
 
     static final Logger LOG = LoggerFactory.getLogger(ModuleController.class);
 
+    @Value("${minio.sbombucket:sboms}")
+    private String bucketName;
+
+    @Value("${sbom.enabled}")
+    private boolean reportGenerationEnabled;
+
     @Autowired
     private ModuleService moduleService;
+
+    @Autowired
+    private MinioService minioService;
 
     @Operation(summary = "Get all module")
     @GetMapping
     public List<ModuleEntity> findAll() {
         return this.moduleService.findAll();
+    }
+
+    @Operation(summary = "true if report generation is enabled")
+    @GetMapping(value = "/reports/enabled")
+    public boolean isEnabled() {
+        return reportGenerationEnabled;
     }
 
     @Operation(summary = "Get all modules with decision count")
@@ -58,7 +80,7 @@ public class ModuleController {
     @Operation(summary = "Get module by name")
     @GetMapping(value = "/byname/{name}")
     public ModuleEntity findByName(@PathVariable("name") String name) {
-        var entityList = this.moduleService.findByName(name);
+        List<ModuleEntity> entityList = this.moduleService.findByName(name);
         if (entityList != null && !entityList.isEmpty()) {
             return this.moduleService.findByName(name).getFirst();
         } else {
@@ -90,5 +112,29 @@ public class ModuleController {
         LOG.info("Module not found. {}", ex.getMessage());
         NotificationDto output = new NotificationDto("error.module.notfound", "Module not found.");
         return new ResponseEntity<>(output, HttpStatus.NOT_FOUND);
+    }
+
+    @GetMapping("/sbom/{id}/{format}")
+    public ResponseEntity<byte[]> downloadSBom(@PathVariable("id") Long id, @PathVariable("format") String format)
+            throws InvalidKeyException, IOException, MinioException {
+
+        ModuleEntity entity = this.moduleService.findById(id);
+        String objectName = entity.getName() + "_" + entity.getVersion() + "." + format;
+
+        // TODO something is wrong in module registry
+        if ("json".equals(format)) {
+            for (String sbomLocation : entity.getSbomlocations().keySet()) {
+                objectName = entity.getName() + "_" + sbomLocation + "_" + entity.getVersion() + "." + format;
+            }
+        }
+
+        byte[] file = minioService.getFileFromMinio(bucketName, objectName);
+        HttpHeaders header = new HttpHeaders();
+        header.add(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=" + objectName);
+        return ResponseEntity.ok()
+                .headers(header)
+                .contentLength(Long.valueOf(file.length))
+                .contentType(MediaType.APPLICATION_OCTET_STREAM)
+                .body(file);
     }
 }
