+++ service/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -49,6 +49,11 @@
             "name": "default.module.name",
             "type": "java.lang.String",
             "description": "If no module is defined - this module is used"
+        },
+        {
+            "name": "minio.sbombucket",
+            "type": "java.lang.String",
+            "description": "A description for 'minio.sbombucket'"
         }
     ]
 }
